Fixed epsilon for sarsa

+++ src/Agent/tetris_agent.jl
@@ -12,7 +12,7 @@ function Base.show(io::IO, agent::AbstractAgent)
     println("record => ", agent.record)
 end
 
-function get_action(agent::AbstractAgent, nb_outputs=7)
+function get_action(agent::AbstractAgent, nb_outputs::Integer=7)
     final_move = zeros(Int, nb_outputs)
     move = rand(1:nb_outputs)
     final_move[move] = 1
